+++ src/options.ts
@@ -11,6 +11,7 @@ export interface Options {
   macroOpcodes?: boolean; // create combined opcodes for repeated instruction sequences?
   specializedOpcodes?: boolean; // create specialized opcodes for commonly used opcode+operand pairs?
   aliasedOpcodes?: boolean; // create duplicate opcodes for commonly used opcodes?
+  antiInstrumentation?: boolean; // add fake opcode effects to hinder opcode instrumentation?
   timingChecks?: boolean; // add timing checks to detect debuggers?
   concealConstants?: boolean; // conceal strings and integers in the constant pool?
   classObfuscation?: boolean; // obfuscate the VM runtime classes?